+++ src/rnn/common.rs
@@ -1,7 +1,9 @@
 pub mod rnn_error;
+pub mod signal;
 pub mod utils;
 
 pub mod spec_type;
 
+pub mod command;
 pub mod input_cfg;
 pub mod network_cfg;