+++ src/rnn/neural/dendrite.rs
@@ -2,7 +2,7 @@ use std::sync::Arc;
 
 use tokio::sync::{broadcast::Receiver, RwLock};
 
-use crate::rnn::common::input_cfg::InputCfg;
+use crate::rnn::common::{input_cfg::InputCfg, signal::Signal};
 
 /// Neuron's input (the dendrite)
 #[derive(Debug)]
@@ -11,11 +11,11 @@ pub struct Dendrite {
     pub config: InputCfg,
 
     /// Synapse capacity
-    pub syn_capacity: u8,
+    pub synapse_capacity: Signal,
 
     /// Keep party id or none
     pub connected: Option<String>,
 
     /// Receiver part of channel between axon and synapse
-    pub synapse: Option<Arc<RwLock<Receiver<u8>>>>,
+    pub synapse: Option<Arc<RwLock<Receiver<Signal>>>>,
 }